% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/artist_get_correction.R,
%   R/artist_get_top_albums.R, R/artist_get_top_tags.R,
%   R/artist_get_top_tracks.R, R/lastrfm_artist.R
\name{artist_get_correction}
\alias{artist_get_correction}
\alias{artist_get_top_albums}
\alias{artist_get_top_tags}
\alias{artist_get_top_tracks}
\alias{lastrfm-artist}
\title{last.fm Artist API methods}
\usage{
artist_get_correction(artist)

artist_get_top_albums(artist, autocorrect = TRUE, n = NULL)

artist_get_top_tags(artist, autocorrect = TRUE, n = NULL)

artist_get_top_tracks(artist, autocorrect = TRUE, n = page_limit)
}
\arguments{
\item{artist}{The artist name (required)}

\item{autocorrect}{Transform misspelled artist names into correct artist
names, returning the correct version instead (optional)}

\item{n}{Number of entries to return (optional)}
}
\value{
Correct artist name.

Data.frame of top \code{n} albums with name, mbid ID, and last.fm url;
  ordered by popularity.

Data.frame of top \code{n} tags with count, name, and url; ordered by
  popularity.

Data.frame of top \code{n} tags with name, play count, number of
  listeners, mbid ID, and last.fm url; ordered by popularity.
}
\description{
These functions return information on artits. None of these services require
authentication, except for \code{artist_add_tags} and
\code{artist_remove_tags}.
}
\details{
If \code{n} is omitted, the functions return:
\itemize{
 \item{All entries for \code{artist_get_top_tags}, \code{artist_get_top_albums}}
 \item{A number of entries up to \code{page_limit} for \code{artist_get_top_tracks}}
}
}
\section{Functions}{
\itemize{
\item \code{artist_get_correction}: Get corrected artist name

\item \code{artist_get_top_albums}: Get artist top albums, ordered by popularity

\item \code{artist_get_top_tags}: Get artist top tags, ordered by popularity

\item \code{artist_get_top_tracks}: Get artist top tracks, ordered by popularity
}}

\examples{
artist_get_correction("Radioheat")
artist_get_top_albums("Pink Floyd", n = 5)
artist_get_top_tags("Mogwai", n = 5)
artist_get_top_tracks("Radiohead", n = 5)
}
