% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lastrfm.r
\docType{package}
\name{lastrfm-package}
\alias{lastrfm}
\alias{lastrfm-package}
\title{lastrfm: Read data from the last.fm APIs}
\description{
Wrapper around the last.fm APIs.
Read data about artists, albums, tracks, users, etc.
The supported methods are documented at http://www.last.fm/api/show/.
Check the index for information about specific functions.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/miccio-dk/lastrfm}
  \item Report bugs at \url{https://github.com/miccio-dk/lastrfm/issues}
}

}
\author{
\strong{Maintainer}: Riccardo Miccini \email{riccardo.miccini@gmail.com}

}
\keyword{internal}
