% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lastrfm_user.R, R/user_get_recent_tracks.R
\name{lastrfm-user}
\alias{lastrfm-user}
\alias{user_get_recent_tracks}
\title{last.fm User API methods}
\usage{
user_get_recent_tracks(user, time_range = c(NA, NA), n = page_limit)
}
\arguments{
\item{user}{The user name to fetch data for (required)}

\item{n}{Number of entries to return (optional)}

\item{period}{The time period over which to retrieve data for (optional,
default is \code{overall})}

\item{artist}{The artist name (TODO)}

\item{album}{The album name (TODO)}
}
\value{
Data.frame of top \code{n} tags with name, play count, number of
  listeners, mbid ID, and last.fm url; ordered by popularity.
}
\description{
These functions return information on users. None of these services require
authentication.
}
\details{
If \code{n} is omitted, the functions return:
\itemize{
 \item{All entries for TODO}
 \item{A number of entries up to \code{page_limit} for TODO}
}

Field \code{period} can be one of \code{overall}, \code{7day}, \code{1month},
\code{3month}, \code{6month}, or \code{12month}.
}
\section{Functions}{
\itemize{
\item \code{user_get_recent_tracks}: Get artist top tracks, ordered by popularity
}}

\examples{
artist_get_top_tracks("Radiohead", n = 5)
}
