% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lrfm_album_get_top_tags.R
\name{lrfm_album_get_top_tags}
\alias{lrfm_album_get_top_tags}
\title{Get album top tags}
\usage{
lrfm_album_get_top_tags(artist, album, autocorrect = TRUE, n = NULL)
}
\arguments{
\item{artist}{The artist name (required)}

\item{album}{The album name (required)}

\item{autocorrect}{Transform misspelled artist names into correct artist names,
returning the correct version instead (optional)}

\item{n}{Number of tags to return (optional, default is all)}
}
\value{
Data.frame of top \code{n} tags with count, name, and url; ordered by popularity.
}
\description{
Get the top tags for an album on Last.fm, ordered by popularity.
This service does not require authentication.
}
\examples{
lrfm_album_get_top_tags("Radiohead", "OK Computer", n = 5)
}
