% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lrfm_artist_get_correction.R
\name{lrfm_artist_get_correction}
\alias{lrfm_artist_get_correction}
\title{Get corrected artist name}
\usage{
lrfm_artist_get_correction(artist)
}
\arguments{
\item{artist}{The artist name to correct (required)}
}
\value{
Correct artist name.
}
\description{
Use the last.fm corrections data to check whether the supplied artist has a correction to a canonical artist
This service does not require authentication.
}
\examples{
lrfm_artist_get_correction("Radioheat")
}
